backend/requirements.txt
langchain-openai
langchain-google-genai
langgraph
pydantic
python-dotenv
jupyter
ipykernel
pandas
langchain-community
langchain
